edge and cluster metrics refactoring
